+++ population.hpp
@@ -11,21 +11,21 @@
 class Population {
 public:
     Population();
-    Population(int, double=0.02);
+    Population(long, double=0.02);
     Population(const Population&);
 
     NeuralNetwork fittest_member();
     long fitness_sum();
     void make_population();
 
-    int generation() const { return this->current_generation_; }
-    void generation(int g) { this->current_generation_ = g; }
+    long generation() const { return this->current_generation_; }
+    void generation(long g) { this->current_generation_ = g; }
 
-    int best_current_generation() const { return this->best_current_generation_; }
-    void best_current_generation(int b) { this->best_current_generation_ = b; }
+    long best_current_generation() const { return this->best_current_generation_; }
+    void best_current_generation(long b) { this->best_current_generation_ = b; }
 
-    int best_overall() const { return this->best_overall_; }
-    void best_overall(int b) { this->best_overall_ = b;}
+    long best_overall() const { return this->best_overall_; }
+    void best_overall(long b) { this->best_overall_ = b;}
 
     size_t size() const { return this->size_; }
     double mutation_rate() const { return this->mutation_rate_; }
@@ -45,9 +45,9 @@ private:
 
     size_t size_;
     double mutation_rate_;
-    int current_generation_;
-    int best_overall_;
-    int best_current_generation_;
+    long current_generation_;
+    long best_overall_;
+    long best_current_generation_;
     std::vector<NeuralNetwork> population;
 };
